Changed block property alteration approach
I've been taking the wrong approach to editing blocks. Replacing the
block in the block registry will cause incompatibility with or crashes
in mods that checks instanceof on blocks or assumes a block type and
does casting. It might even conflict with Minecraft itself for all I
know. Instead, I'll modify the blocks already in the block registry;
hopefully this will work better with other mods and the game itself.